// packages/mavlink/src/generated/messages.ts
export const HEARTBEAT: MessageDescriptor<'HEARTBEAT'> = {
  id: 0,
  name: 'HEARTBEAT',
  crcExtra: 50,
  baseLength: 9,
  length: 9,
  fields: [
    { name: 'type', type: 'uint8_t', offset: 4, enum: MavType },
    { name: 'autopilot', type: 'uint8_t', offset: 5, enum: MavAutopilot },
    { name: 'baseMode', type: 'uint8_t', offset: 6, enum: MavModeFlag, bitmask: true },
    { name: 'customMode', type: 'uint32_t', offset: 0 },
    { name: 'systemStatus', type: 'uint8_t', offset: 7, enum: MavState },
    { name: 'mavlinkVersion', type: 'uint8_t', offset: 8 }
  ]
}
// ...
export const SYS_STATUS: MessageDescriptor<'SYS_STATUS'> = {
  id: 1,
  name: 'SYS_STATUS',
  crcExtra: 124,
  baseLength: 31,
  length: 43,
  fields: [
    { name: 'onboardControlSensorsPresent', type: 'uint32_t', offset: 0, enum: MavSysStatusSensor, bitmask: true },
    { name: 'onboardControlSensorsEnabled', type: 'uint32_t', offset: 4, enum: MavSysStatusSensor, bitmask: true },
    { name: 'onboardControlSensorsHealth', type: 'uint32_t', offset: 8, enum: MavSysStatusSensor, bitmask: true },
    { name: 'load', type: 'uint16_t', offset: 12 },
    { name: 'voltageBattery', type: 'uint16_t', offset: 14 },
    { name: 'currentBattery', type: 'int16_t', offset: 16 },
    { name: 'batteryRemaining', type: 'int8_t', offset: 30 },
    { name: 'dropRateComm', type: 'uint16_t', offset: 18 },
    { name: 'errorsComm', type: 'uint16_t', offset: 20 },
    { name: 'errorsCount1', type: 'uint16_t', offset: 22 },
    { name: 'errorsCount2', type: 'uint16_t', offset: 24 },
    { name: 'errorsCount3', type: 'uint16_t', offset: 26 },
    { name: 'errorsCount4', type: 'uint16_t', offset: 28 },
    {
      name: 'onboardControlSensorsPresentExtended',
      type: 'uint32_t',
      offset: 31,
      extension: true,
      enum: MavSysStatusSensorExtended,
      bitmask: true
    },
    {
      name: 'onboardControlSensorsEnabledExtended',
      type: 'uint32_t',
      offset: 35,
      extension: true,
      enum: MavSysStatusSensorExtended,
      bitmask: true
    },
    {
      name: 'onboardControlSensorsHealthExtended',
      type: 'uint32_t',
      offset: 39,
      extension: true,
      enum: MavSysStatusSensorExtended,
      bitmask: true
    }
  ]
}
// ...
export const ATTITUDE: MessageDescriptor<'ATTITUDE'> = {
  id: 30,
  name: 'ATTITUDE',
  crcExtra: 39,
  baseLength: 28,
  length: 28,
  fields: [
    { name: 'timeBootMs', type: 'uint32_t', offset: 0 },
    { name: 'roll', type: 'float', offset: 4 },
    { name: 'pitch', type: 'float', offset: 8 },
    { name: 'yaw', type: 'float', offset: 12 },
    { name: 'rollspeed', type: 'float', offset: 16 },
    { name: 'pitchspeed', type: 'float', offset: 20 },
    { name: 'yawspeed', type: 'float', offset: 24 }
  ]
}
// ...
export const MISSION_CLEAR_ALL: MessageDescriptor<'MISSION_CLEAR_ALL'> = {
  id: 45,
  name: 'MISSION_CLEAR_ALL',
  crcExtra: 232,
  baseLength: 2,
  length: 3,
  fields: [
    { name: 'targetSystem', type: 'uint8_t', offset: 0 },
    { name: 'targetComponent', type: 'uint8_t', offset: 1 },
    { name: 'missionType', type: 'uint8_t', offset: 2, extension: true, enum: MavMissionType }
  ]
}